refactor builder config (env => target)
